chore: remove top-level types, update imports, and format API reference
