+++ apps/web/vibes/soul/brands.ts
@@ -128,4 +128,6 @@ export const brands = [
       '--shadow-xl': '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
     },
   },
-] satisfies Brands
+] as const satisfies Brands
+
+export type SoulBrandName = (typeof brands)[number]['name']
