+++ apps/web/vibes/soul/docs/products-list.mdx
@@ -41,12 +41,15 @@ The `products` property accepts either `Product[]` or `Promise<Product[]>`. By p
 
 ### `ProductsList`
 
-| Prop       | Type                                                                                                      | Default |
-| ---------- | --------------------------------------------------------------------------------------------------------- | ------- |
-| `products` | [`Product[]`](/docs/soul/product-card#product) \| [`Promise<Product[]>`](/docs/soul/product-card#product) |         |
+| Prop                | Type                                                                                                                         | Default |
+| ------------------- | ---------------------------------------------------------------------------------------------------------------------------- | ------- |
+| `products`          | [`Product`](/docs/soul/product-card#product)`[]` &nbsp;\| &nbsp;`Promise<`[`Product>`](/docs/soul/product-card#product)`[]>` |         |
+| `className`         | `string`                                                                                                                     |         |
+| `showCompare?`      | `boolean`                                                                                                                    | `false` |
+| `compareProducts?`  | [`Product`](/docs/soul/product-card#product)`[]` &nbsp;\|&nbsp; `Promise<`[`Product`](/docs/soul/product-card#product)`[]>`  |         |
+| `compareLabel?`     | `string`                                                                                                                     |         |
+| `compareParamName?` | `string`                                                                                                                     |         |
 
 ## Figma
-
-TODO - need figma link
 
 <Figma url="https://www.figma.com/design/kgCf8OkDCe90wBSKNNqueL/Soul-Vibe?node-id=1720-21795&t=RzBeIql49iCnmSMw-4" />
