+++ apps/web/vibes/soul/primitives/product-chip/index.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import { clsx } from 'clsx'
 import { X } from 'lucide-react'
 
-import { ProductCardProduct } from '../../types'
+import { ProductCardProduct } from '@/vibes/soul/primitives/product-card'
 
 function getInitials(name: string): string {
   return name
