+++ apps/web/vibes/soul/sections/featured-products-list/index.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link'
 
 import { Button } from '@/vibes/soul/primitives/button'
+import { ProductCardProduct } from '@/vibes/soul/primitives/product-card'
 import { ProductsList } from '@/vibes/soul/primitives/products-list'
-import { ProductCardProduct } from '@/vibes/soul/types'
 
 interface Link {
   label: string
