+++ apps/web/vibes/soul/sections/products-list-section/mobile-filters.tsx
@@ -2,9 +2,8 @@ import { Sliders } from 'lucide-react'
 
 import { Button } from '@/vibes/soul/primitives/button'
 import * as SidePanel from '@/vibes/soul/primitives/side-panel'
-import { Filter } from '@/vibes/soul/types'
 
-import { FiltersPanel } from './filters-panel'
+import { Filter, FiltersPanel } from './filters-panel'
 
 interface Props {
   filters: Filter[] | Promise<Filter[]>
