+++ apps/web/vibes/soul/sections/products-list-section/products-list-section.tsx
@@ -1,18 +1,18 @@
-import { Breadcrumbs } from '@/vibes/soul/primitives/breadcrumbs'
+import { Breadcrumb, Breadcrumbs } from '@/vibes/soul/primitives/breadcrumbs'
+import { ProductCardProduct } from '@/vibes/soul/primitives/product-card'
 import { ProductsList } from '@/vibes/soul/primitives/products-list'
-import {
-  Breadcrumb,
-  Filter,
-  Pages,
-  ProductCardProduct,
-  Option as SortOption,
-} from '@/vibes/soul/types'
 
-import { FiltersPanel } from './filters-panel'
+import { Filter, FiltersPanel } from './filters-panel'
 import { MobileFilters } from './mobile-filters'
 import { Pagination, PaginationInfo } from './pagination'
+import { Option as SortOption } from './sorting'
 import { Sorting } from './sorting'
 
+interface Pages {
+  name: string
+  previousPage?: string
+  nextPage?: string
+}
 interface Props {
   breadcrumbs?: Breadcrumb[]
   title?: Promise<string> | string
