--- apps/web/vibes/soul/types/breadcrumbs.ts
@@ -1,4 +1,0 @@
-export interface Breadcrumb {
-  label: string
-  href: string
-}
