--- apps/web/vibes/soul/types/image.ts
@@ -1,4 +1,0 @@
-export interface Image {
-  alt: string
-  src: string
-}
