--- apps/web/vibes/soul/types/index.ts
@@ -1,13 +1,0 @@
-import { brands } from '@/vibes/soul/brands'
-
-// const brandNames = brands.map(brand => brand.name)
-// export type Brand = (typeof brandNames)[number]
-
-export type Brand = 'Electric' | 'Warm' | 'Luxury'
-
-export * from './cart'
-export * from './filter'
-export * from './image'
-export * from './product'
-export * from './breadcrumbs'
-export * from './pagination'
