--- apps/web/vibes/soul/types/pagination.ts
@@ -1,5 +1,0 @@
-export interface Pages {
-  name: string
-  previousPage?: string
-  nextPage?: string
-}
