--- apps/web/vibes/soul/types/product.ts
@@ -1,31 +1,0 @@
-import { Image } from './image'
-
-export interface ProductCardProduct {
-  id: string
-  title: string
-  href: string
-  image?: Image
-  price?: ProductCardPrice
-  subtitle?: string
-  badge?: string
-  rating?: number
-}
-
-export interface CompareCardProduct extends ProductCardProduct {
-  description?: string
-  availability?: string
-}
-
-export interface ProductCardPriceRange {
-  type: 'range'
-  minValue: string
-  maxValue: string
-}
-
-export type ProductCardPrice = string | ProductCardPriceRange | ProductCardSalePrice
-
-export interface ProductCardSalePrice {
-  type: 'sale'
-  previousValue: string
-  currentValue: string
-}
